ktu-pyq-analyzer/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.9
Pillow==10.3.0
numpy<2
pypdf==4.2.0
pdf2image==1.17.0
python-jose[cryptography]==3.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
aiofiles==23.2.1
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
